+++ auto_factory/example/lib/src/application.dart
@@ -1,17 +1,23 @@
 import 'package:auto_factory_annotation/auto_factory_annotation.dart';
 
+import 'provider/provider.dart';
 import 'service/hello_service.dart';
 
 part 'application.g.dart';
 
-@AutoFactory()
+@Injectable()
 class Application {
-  HelloService helloService;
+  final HelloService helloService;
+  final String providedName;
 
-  Application(this.helloService);
+  Application(
+    this.helloService,
+    @ProvidedBy(HelloServiceProvider) this.providedName,
+  );
 
   void greet() {
     final name = helloService.who('John Doe');
     print('Hello $name');
+    print('Hello $providedName');
   }
 }