+++ auto_factory/example/lib/src/service/hello_service.dart
@@ -2,7 +2,7 @@ import 'package:auto_factory_annotation/auto_factory_annotation.dart';
 
 part 'hello_service.g.dart';
 
-@CachingFactory()
+@Singleton()
 class HelloService {
   String who(String name) {
     return name ?? 'World';