+++ auto_factory/lib/src/provider_generator.dart
@@ -10,6 +10,8 @@ import 'package:source_gen/source_gen.dart';
  * Base class extended by [FactoryGenerator] and [CachingFactoryGenerator] 
  */
 abstract class BaseFactoryGenerator<T> extends GeneratorForAnnotation<T> {
+  final TypeChecker _providedChecker = TypeChecker.fromRuntime(ProvidedBy);
+
   @override
   FutureOr<String> generateForAnnotatedElement(
     Element element,
@@ -17,24 +19,50 @@ abstract class BaseFactoryGenerator<T> extends GeneratorForAnnotation<T> {
     BuildStep buildStep,
   ) {
     if (element is ClassElement) {
-      final ctor =
-          element.constructors.firstWhere((ctor) => ctor.displayName == '');
-
-      final declarations = ctor.parameters
-          .map((param) => param.type)
-          .map(_declaration)
+      final declarations = element.unnamedConstructor.parameters
+          .map(_generateDeclaration)
           .join('\n');
 
-      final params =
-          ctor.parameters.map((param) => param.type).map(_parameter).join(' ');
+      final params = element.unnamedConstructor.parameters
+          .map(_generateParamater)
+          .join(' ');
 
       return _finalize(element, declarations, params);
     }
     return '';
   }
 
-  String _declaration(DartType type);
-  String _parameter(DartType type);
+  String _generateDeclaration(ParameterElement element) {
+    final neededType = _providedChecker.hasAnnotationOfExact(element)
+        ? _providedChecker
+            .firstAnnotationOfExact(element)
+            .getField('provider')
+            .toTypeValue()
+        : element.type;
+
+    return 'final ${_typeToInstanceName(neededType)}Factory = ${_typeToString(neededType)}Factory();';
+  }
+
+  String _generateParamater(ParameterElement element) {
+    if (_providedChecker.hasAnnotationOfExact(element)) {
+      final elementType = _providedChecker
+          .firstAnnotationOfExact(element)
+          .getField('provider')
+          .toTypeValue();
+      return 'await (await ${_typeToInstanceName(elementType)}Factory.create()).provide(),';
+    }
+    return 'await ${_typeToInstanceName(element.type)}Factory.create(),';
+  }
+
+  String _typeToString(DartType type) {
+    return type.element.name;
+  }
+
+  String _typeToInstanceName(DartType type) {
+    final typeAsString = type.element.name;
+    return typeAsString[0].toLowerCase() + typeAsString.substring(1);
+  }
+
   String _finalize(
     ClassElement element,
     String declarations,
@@ -42,23 +70,46 @@ abstract class BaseFactoryGenerator<T> extends GeneratorForAnnotation<T> {
   );
 }
 
-/// Generates a factory shared part for classes annotated with [AutoFactory].
+/// Generates a provider implementation for classes annotated with [Provider].
 /// After creating an instance, the method get provides the required type and its dependencies
-class FactoryGenerator extends BaseFactoryGenerator<AutoFactory> {
-  @override
-  String _declaration(DartType type) =>
-      'final ${_typeToInstanceName(type)}Factory = ${_typeToString(type)}Factory();';
-
-  @override
-  String _parameter(DartType type) =>
-      '${_typeToInstanceName(type)}Factory.get(),';
-
+class ProviderGenerator extends BaseFactoryGenerator<Provider> {
   @override
   String _finalize(ClassElement element, String declarations, String params) =>
       '''
       class ${element.name}Factory {
 
-        ${element.name} get() {
+        static final ${element.name}Factory _singleton = ${element.name}Factory._internal();
+
+        factory ${element.name}Factory() {
+          return _singleton;
+        }
+
+        ${element.name}Factory._internal();
+
+        ${element.name} _objectInstance;
+
+        Future<${element.name}> create() async {
+          $declarations
+
+          _objectInstance ??= ${element.name}(
+            $params
+          );
+
+          return _objectInstance;
+        }
+      }
+      ''';
+}
+
+/// Generates a factory shared part for classes annotated with [Injectable].
+/// After creating an instance, the method create provides the required type and its dependencies
+class FactoryGenerator extends BaseFactoryGenerator<Injectable> {
+  @override
+  String _finalize(ClassElement element, String declarations, String params) =>
+      '''
+      class ${element.name}Factory {
+
+        Future<${element.name}> create() async {
 
           $declarations
 
@@ -68,43 +119,33 @@ class FactoryGenerator extends BaseFactoryGenerator<AutoFactory> {
       ''';
 }
 
-/// Generates a caching factory shared part for classes annotated with [CachingFactory].
-/// This is the same as [AutoFactory] but the instance is cached
-class CachingFactoryGenerator extends BaseFactoryGenerator<CachingFactory> {
-  @override
-  String _declaration(DartType type) =>
-      'final ${_typeToInstanceName(type)}Factory = ${_typeToString(type)}Factory();';
-
-  @override
-  String _parameter(DartType type) {
-    return '${_typeToInstanceName(type)}Factory.get(),';
-  }
-
+/// Generates a caching factory shared part for classes annotated with [Singleton].
+/// This is the same as [Injectable] but the instance is cached
+class CachingFactoryGenerator extends BaseFactoryGenerator<Singleton> {
   @override
   String _finalize(ClassElement element, String declarations, String params) =>
       '''
       class ${element.name}Factory {
 
-        ${element.name} instance;
+        static final ${element.name}Factory _singleton = ${element.name}Factory._internal();
 
-        ${element.name} get() {
+        factory ${element.name}Factory() {
+          return _singleton;
+        }
+
+        ${element.name}Factory._internal();
+
+        ${element.name} _objectInstance;
+
+        Future<${element.name}> create() async {
           $declarations
 
-          instance ??= ${element.name}(
+          _objectInstance ??= ${element.name}(
             $params
           );
 
-          return instance;
+          return _objectInstance;
         }
       }
       ''';
-}
-
-String _typeToString(DartType type) {
-  return type.element.name;
-}
-
-String _typeToInstanceName(DartType type) {
-  final typeAsString = type.element.name;
-  return typeAsString[0].toLowerCase() + typeAsString.substring(1);
 }