+++ auto_factory/test/sources/_basic_fourth_dep.dart
@@ -1,0 +1,6 @@
+import 'package:auto_factory_annotation/auto_factory_annotation.dart';
+
+@Injectable()
+class FourthDep {
+  String name = 'fourth_dep';
+}