+++ auto_factory/test/sources/_caching.dart
@@ -2,27 +2,41 @@ import 'package:auto_factory_annotation/auto_factory_annotation.dart';
 import 'package:source_gen_test/annotations.dart';
 
 import '_basic_dep.dart';
+import '_basic_provider.dart';
+import '_basic_third_dep.dart';
 
 @ShouldGenerate('''
 class MyComponentFactory {
-  MyComponent instance;
+  static final MyComponentFactory _singleton = MyComponentFactory._internal();
 
-  MyComponent get() {
+  factory MyComponentFactory() {
+    return _singleton;
+  }
+
+  MyComponentFactory._internal();
+
+  MyComponent _objectInstance;
+
+  Future<MyComponent> create() async {
     final dependencyFactory = DependencyFactory();
+    final thirdDependencyProviderFactory = ThirdDependencyProviderFactory();
 
-    instance ??= MyComponent(
-      dependencyFactory.get(),
+    _objectInstance ??= MyComponent(
+      await dependencyFactory.create(),
+      await (await thirdDependencyProviderFactory.create()).provide(),
     );
 
-    return instance;
+    return _objectInstance;
   }
 }
 ''')
-@CachingFactory()
+@Singleton()
 class MyComponent {
-  final Dependency _dep;
+  final Dependency dep;
+  final ThirdDependency third;
 
-  MyComponent(Dependency dep) : _dep = dep;
-
-  Dependency get dep => _dep;
+  MyComponent(
+    this.dep,
+    @ProvidedBy(ThirdDependencyProvider) this.third,
+  );
 }